More readable output & introduce bpaf for CLI parsing
Started with the goal of having a more readable output.
This is accomplished through introducing a verbose flag (`-v`).
Verbosity mode is disabled by default, and by default during transfer,
a single line of output is shown such as:
```
[0.05 GB / 118.45 GB] 0.0% 6.00 MB/s, 5h 28m left
```
which refreshes as the transfer continues.
This makes the output more readable.

For debugging, `-v` can still be passed and will output the verbsoe
output as previously.

The new verbosity argument was added, and this made the cli
parsing more complex than previously. It now has to consider passing this
flag on different places in the argument order.

To deal with this CLI complexity, introduce `bpaf` cli parsing library
and migrate the current custom CLI parsing implementation to use this library.